Updated 'Dokkaebi Official Post' to 'Dokkaebi Gamtu' (#1370)
In the original Korean text, it calls this term the Dokkaebi Gamtu (도깨비 감투). The translator translated 감투 as Official Post which doesn't make sense, especially considering that Dokja refers to it as an article of clothing in folklore. So, we could translate 도깨비 감투 as 'Dokkaebi Cap' (which is what the ORV wiki calls it), or translate it as Gamtu and add a note.